Function for comparing models

Function compare_models enables user to compare various model with the same type of task. Function compares them based on metric which can be chosen by the user and returns the best model

<<<<<<< Updated upstream
compare_models(models, metric)
=======
compare_models(..., data_test, target, metric = NULL)
>>>>>>> Stashed changes
Arguments

<<<<<<< Updated upstream 	models
list of models to compare. Models shoud be either forester_model or explainer object.
======= 	data_test
data frame for evaluation models

target
character indicating the target column in data test table
>>>>>>> Stashed changes

metric
string containing name of mertic based on which the model will be selected. Metric shoud be written in small letters

models
list of models to compare. Models shoud be either forester_model or explainer object.

Value

A list containing an object of the class forester_model and data frame with results.

References

Explanatory Model Analysis. Explore, Explain, and Examine Predictive Models. https://ema.drwhy.ai/

Examples

# \donttest{
library(DALEX)

#> Welcome to DALEX (version: 2.2.1).
#> Find examples and detailed introduction at: http://ema.drwhy.ai/
#> Additional features will be available after installation of: ggpubr.
#> Use 'install_dependencies()' to get all suggested dependencies

data(apartments, package="DALEX")

lightgbm_explained <- make_lightgbm(apartments, "m2.price", "regression")
<<<<<<< Updated upstream
ranger_explained <- make_ranger(apartments, "m2.price", "regression")

models <- list(lightgbm_explained, ranger_explained)

compare_models(models, "rmse")

#> 
#> 
#> |model    |      rmse|       mse|        r2|      mad|
#> |:--------|---------:|---------:|---------:|--------:|
#> |LightGBM |  84.37347|  7118.883| 0.9913318| 52.30193|
#> |Ranger   | 145.08036| 21048.310| 0.9743710| 82.04869|

#> [[1]]
#> Model label:  LightGBM 
#> Model class:  lgb.Booster,R6 
#> Data head  :
#>   construction.year surface floor no.rooms    district
#> 1              1953      25     3        1 Srodmiescie
#> 2              1992     143     9        5     Bielany
#> 

# }
=======

#> Original shape of train data frame: 1000 rows, 6 columns

#> _____________

#> NA values

#> There is no NA values in your data.

ranger_explained <- make_ranger(apartments, "m2.price", "regression")

#> Original shape of train data frame: 1000 rows, 6 columns

#> _____________

#> NA values

#> There is no NA values in your data.

#> __________________________

#> CREATING MODEL

best_model <- compare_models(lightgbm_explained, ranger_explained, 
                             data_test = apartments, 
                             target = "m2.price",
                             metric = "rmse")

#> [1] "done"
#> [1] "done"

#> Results of compared models:

#> 
#> 
#> model            rmse         mse          r2        mad
#> ---------  ----------  ----------  ----------  ---------
#> LightGBM     83.89545    7038.447   0.9914298   46.31874
#> Ranger      147.97299   21896.006   0.9733388   85.22582

#> The best model based on rmse metric is LightGBM.

# }
>>>>>>> Stashed changes