Modified COMPAS dataset

compas dataset. From ProPublica: across the nation, judges, probation and parole officers are increasingly using algorithms to assess a criminal defendant’s likelihood to re-offend.

data(compas)

Format

A data frame with 6172 rows and 7 variables:

Source

The original source of data is https://www.propublica.org/datastore/dataset/compas-recidivism-risk-score-data-and-analysis. Modified data used here comes from https://www.kaggle.com/danofer/compass/ (probublicaCompassRecidivism_data_fairml.csv)

Details

Two_yr_Recidivism

factor, 1/0 for future recidivism or no recidivism. Models should predict this values

Number_of_Priors

numeric, number of priors

Age_Above_FourtyFive

factor, 1/0 for age above 45 years or not

Age_Below_TwentyFive

factor, 1/0 for age below 25 years or not

Misdemeanor

factor, 1/0 for having recorded misdemeanor(s) or not

Ethnicity

factor, Caucasian, African American, Asian, Hispanic, Native American or Other

Sex

factor, female/male for gender