Function for evaluating and comparing models

Function evaluate enables user to compare one or several models with the same type of task. Function compares them based on metric which can be chosen by the user and returns the best model

evaluate(..., data_test, target, metric = NULL)

Arguments

...
models to compare. Models should be either forester_model or explainer object.

data_test
data frame for evaluation models.

target
character indicating the target column in data test table.

metric
string containing name of metric based on which the model will be selected. Metric should be written in lowercase letters. Can be omitted.

Value

A list containing an object of the class forester_model and data frame with results.

References

Explanatory Model Analysis. Explore, Explain, and Examine Predictive Models. https://ema.drwhy.ai/

Examples

# \donttest{
library(DALEX)

#> Welcome to DALEX (version: 2.2.1).
#> Find examples and detailed introduction at: http://ema.drwhy.ai/
#> Additional features will be available after installation of: ggpubr.
#> Use 'install_dependencies()' to get all suggested dependencies

data(apartments, package="DALEX")
lightgbm_model <- make_lightgbm(apartments, "m2.price", "regression")

#> --- Creating LightGBM model ---

#> Original shape of train data frame: 1000 rows, 6 columns

#> _____________

#> NA values

#> There is no NA values in your data.

ranger_model <- make_ranger(apartments, "m2.price", "regression")

#> --- Creating Ranger model ---

#> Original shape of train data frame: 1000 rows, 6 columns

#> _____________

#> NA values

#> There is no NA values in your data.

# Evaluating single model
model_evaluated <- evaluate(ranger_model, 
                             data_test = apartments, 
                             target = "m2.price",
                             metric = "mse")

#> Results of compared models:

#> 
#> 
#> model          mse      rmse          r2        mad
#> -------  ---------  --------  ----------  ---------
#> Ranger    21896.01   147.973   0.9733388   85.22582

#> The best model based on mse metric is Ranger.

best_model <- evaluate(lightgbm_model, ranger_model, 
                             data_test = apartments, 
                             target = "m2.price",
                             metric = "rmse")

#> Results of compared models:

#> 
#> 
#> model            rmse         mse          r2        mad
#> ---------  ----------  ----------  ----------  ---------
#> LightGBM     83.89545    7038.447   0.9914298   46.31874
#> Ranger      147.97299   21896.006   0.9733388   85.22582

#> The best model based on rmse metric is LightGBM.

# }