Automated Machine Learning Model Solver

Different tree-based models such as: XGBoost, ranger, CatBoost, LightGBM, etc. require different syntaxes and different specific data objects. This function provides a simple and unified formula to create those models with options to automatically cover the whole process of creating Machine Learning Model: Preprocessing Data, Feature Engineering, Creating Models, Optimizing Hyperparameters, Model Explanation and Evaluating Models.

forester(
  data,
  target,
  type,
  metric = NULL,
  data_test = NULL,
  fill_na = TRUE,
  num_features = NULL,
  tune = FALSE,
  tune_iter = 20
)

Arguments

data
data.frame, matrix, data.table or dgCMatrix - training data set to create model, if data_test = NULL, then data will be automatically divided into training and testing dataset. NOTE: data has to contain the target column.

target
character: name of the target column, should be character and has to be column name in data.

type
character: defining the task. Two options are: "regression" and "classification", particularly, binary classification.

metric
character, name of metric used for evaluating best model. For regression, options are: "mse", "rmse", "mad" and "r2". For classification, options are: "auc", "recall", "precision", "f1" and "accuracy".

data_test
optional argument, class of data.frame, matrix, data.table or dgCMatrix - test data set used for evaluating model performance.

fill_na
logical, default is FALSE. If TRUE, missing values in target column are removed, missing values in categorical columns are replaced by mode and missing values in numeric columns are substituted by median of corresponding columns.

num_features
numeric, default is NULL. Parameter indicates number of most important features, which are chosen from the train dataset. Automatically, those important features will be kept in the train and test datasets.

tune
logical. If TRUE, function will perform the hyperparameter tuning steps for each model inside.

tune_iter
number (default: 20) - total number of times the optimization step is to repeated. This argument is used when tune = TRUE.

Value

An object of the class forester_model which is the best model with respect to the chosen metric. It's also an object of the class explainer from DALEX family inherited the explanation for the best chosen model.

Examples

# \donttest{
# regression
library(DALEX)
data(apartments, package="DALEX")

exp <- forester(apartments, "m2.price", "regression")

#> __________________________

#> FORESTER

#> Original shape of train data frame: 800 rows, 6 columns

#> _____________

#> NA values

#> There is no NA values in your data.

#> __________________________

#> CREATING MODELS

#> --- Ranger model has been created ---

#> --- Catboost model has been created ---

#> --- Xgboost model has been created ---

#> --- LightGBM model has been created ---

#> __________________________

#> COMPARISON

#> Results of compared models:

#> 
#> 
#> model            rmse         mse          r2        mad
#> ---------  ----------  ----------  ----------  ---------
#> CatBoost     33.27786    1107.416   0.9986369   17.88165
#> XGBoost      58.52602    3425.295   0.9957839   36.56201
#> LightGBM     91.74899    8417.878   0.9896386   48.94833
#> Ranger      168.81752   28499.356   0.9649208   95.91445

#> The best model based on rmse metric is CatBoost.

# plot(model_parts(exp))

# binary classification
library(DALEX)
data(titanic_imputed, package="DALEX")

exp <- forester(titanic_imputed, "survived", "classification")

#> __________________________

#> FORESTER

#> Original shape of train data frame: 1764 rows, 8 columns

#> _____________

#> NA values

#> There is no NA values in your data.

#> __________________________

#> CREATING MODELS

#> --- Ranger model has been created ---

#> --- Catboost model has been created ---

#> --- Xgboost model has been created ---

#> --- LightGBM model has been created ---

#> __________________________

#> COMPARISON

#> Results of compared models:

#> 
#> 
#> model             auc      recall   precision          f1    accuracy
#> ---------  ----------  ----------  ----------  ----------  ----------
#> LightGBM    0.7872147   0.6549296   0.8711944   0.7477387   0.8577098
#> XGBoost     0.7851539   0.6338028   0.9068010   0.7461140   0.8611111
#> CatBoost    0.7843472   0.6408451   0.8834951   0.7428571   0.8571429
#> Ranger      0.7789845   0.6056338   0.9124668   0.7280423   0.8543084

#> The best model based on auc metric is LightGBM.

# plot(model_parts(exp))
# }