Lisbon dataset

lisbon Explore the regression algorithm using the prices of Lisbon's houses. This dataset contains a total of 246 records.

data(lisbon)

Format

A data frame with 32561 rows and 15 variables:

Id

integer, is a unique identifying number assigned to each house.

Condition

character, The house condition (i.e., New, Used, As New, For Refurbishment).

PropertyType

character, Property type (i.e., Home, Single habitation)

PropertySubType

character, Property Sub Type (i.e., Apartment, duplex, etc.)

Bedrooms

integer, Number of Bedrooms

Bathrooms

integer, Number of Bathrooms

AreaNet

integer, Net area of the house

AreaGross

integer, Gross area of the house

Parking

integer, Number of parking places

Latitude

numeric, Geographical Latitude

Longitude

numeric, Geographical Longitude

Country

character, Country where the house is located

District

character, District where the house is located

Municipality

character, Municipality where the house is located

Parish

character, Parish where the house is located

Price.M2

integer, Price per m² in the location of the house

Price

integer, This is our training variable and target. It is the home price.

Source

Data from Kaggle https://www.kaggle.com/datasets/cgrodrigues/lisbon-house-prices