Perform predictions on new data

This function is used when the user created the models and evaluated them on a datset x and later on, he got new observations and wants to predict the target value for them. The `predict_new()` function quickly normalizes their format, like a subset of columns that were deleted for the models and makes predictions. It is strongly advised that the new dataset shouldn't have any missing values.

predict_new(train_out, data, verbose = TRUE)

Arguments

train_out

The output of the `train()` function.

data

A dataset with the same format as the one used in `train()` function, but with new observations.

verbose

A logical value, if set to TRUE, provides all information about the process, if FALSE gives none.

Value

A list of predictions.

Examples

data(lisbon)
lisbon_train <- lisbon[1:200, ]
lisbon_new   <- lisbon[201:246, ]
out          <- train(lisbon_train, 'Price', verbose = FALSE)
#> Warning: Number of logged events: 4
#> Warning: Number of logged events: 1
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
preds        <- predict_new(out, lisbon_new)
#> Warning: Number of logged events: 2
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
preds
#> [[1]]
#>  [1]  952324.4  344492.9  339461.3  615496.3  419187.0 1061632.1  311485.4
#>  [8]  516179.9  310106.2  334668.3  910501.2  340784.6  311194.8  491818.4
#> [15]  632836.6  591809.5  509900.9  378525.7  443603.2  445102.4  460494.7
#> [22]  502513.4  519582.9  311231.8  820227.6  532248.6  346908.5  332178.5
#> [29]  315635.9  327569.1  301972.0  495371.2  473809.9  468653.7  448073.5
#> [36]  943346.2  366005.6  880292.2  344630.9  434296.8  336840.7  345336.5
#> [43]  491818.4  444452.4  723162.4  433400.6
#> 
#> [[2]]
#>  [1]  704335.6  276345.3  241238.8  639040.9  301120.6  815301.8  199242.4
#>  [8]  473290.3  239686.5  262934.0 1174809.5  279636.9  218607.9  387491.8
#> [15]  639040.9  449130.5  492685.2  397723.6  365181.2  312910.1  460595.4
#> [22]  473290.3  493889.1  242669.7  795324.6  471143.6  247203.8  274959.5
#> [29]  253493.6  276982.1  239686.5  397446.7  397446.7  394791.9  397637.8
#> [36] 1034775.6  251675.0  941044.5  293526.9  395172.9  288723.2  229590.7
#> [43]  387491.8  369891.2  596814.4  398149.2
#> 
#> [[3]]
#>  [1] 785583.3 278263.4 278263.4 471314.3 278263.4 785583.3 278263.4 471314.3
#>  [9] 278263.4 278263.4 785583.3 278263.4 278263.4 278263.4 471314.3 471314.3
#> [17] 278263.4 278263.4 278263.4 471314.3 471314.3 471314.3 278263.4 278263.4
#> [25] 785583.3 471314.3 278263.4 278263.4 278263.4 278263.4 278263.4 278263.4
#> [33] 278263.4 278263.4 278263.4 785583.3 471314.3 785583.3 278263.4 471314.3
#> [41] 278263.4 278263.4 278263.4 278263.4 471314.3 471314.3
#> 
#> [[4]]
#>  [1]  897514.6  283736.8  272127.7  921523.4  345044.2  652307.9  141502.9
#>  [8]  453432.2  197132.2  272231.5 1023169.4  283736.8  236665.3  425839.7
#> [15]  921523.4  648266.5  501597.8  225451.5  391738.8  298369.5  402356.3
#> [22]  412057.7  467412.9  236665.3  580505.1  435168.3  284556.5  230918.0
#> [29]  316127.9  283736.8  207218.1  426140.2  388914.4  412747.2  359709.6
#> [36]  957915.0  226624.9  895412.3  260466.9  330809.0  239939.6  238572.2
#> [43]  425839.7  390415.5  725291.8  242750.5
#> 
#> [[5]]
#>  [1]  753047.2  299638.2  263288.5  668645.1  350881.6  821294.3  229385.0
#>  [8]  443276.8  226271.9  248967.3 1272516.7  316379.0  263307.5  457115.5
#> [15]  737411.3  578404.3  494023.3  295780.6  423088.0  391547.2  423687.2
#> [22]  445306.2  501976.2  263627.1  641471.9  480160.6  280799.6  259590.8
#> [29]  270338.5  272699.5  223266.8  452169.6  410595.6  397093.7  388041.4
#> [36] 1201340.6  323533.7  740961.4  308523.2  369002.3  243093.6  271742.4
#> [43]  457115.5  407819.6  597531.6  357554.4
#> 
#> [[6]]
#>  [1]  875490.8  343850.6  345290.7  583635.9  402437.7 1029565.7  287822.7
#>  [8]  501849.2  301781.1  320336.4  882514.1  347065.1  277337.9  469330.7
#> [15]  621279.8  603267.0  476083.9  354867.1  440265.4  432748.2  483956.3
#> [22]  487430.9  492536.3  279177.7  827793.2  499431.3  316791.2  304404.3
#> [29]  303160.4  329105.2  276918.6  481777.2  440623.0  453394.9  417384.1
#> [36]  907087.7  355852.1  878886.1  322051.4  425747.4  311520.2  354983.0
#> [43]  469330.7  465703.4  662189.5  418989.8
#> 
#> [[7]]
#>  [1]  835800.0  330180.8  333738.3  575064.5  396393.1 1086844.6  278469.1
#>  [8]  500200.5  279491.2  320697.7  899725.0  330051.8  268717.6  468479.4
#> [15]  596664.5  548335.8  487534.8  350048.5  431533.9  439862.8  462000.8
#> [22]  498205.5  497001.4  273611.6  796732.0  492235.9  294238.3  287639.7
#> [29]  287673.7  302418.3  271137.8  460718.5  420858.5  407800.8  382665.8
#> [36]  952917.5  328818.0  823449.6  324041.5  405710.0  288728.8  309925.3
#> [43]  468479.4  446504.2  586956.7  401770.5
#> 
#> [[8]]
#>  [1] 881857.4 359571.9 359711.4 565742.1 397211.8 904648.5 307039.9 503089.2
#>  [9] 282426.3 302957.0 861192.5 357452.4 320274.2 494308.3 573719.4 598526.3
#> [17] 502863.2 389140.2 443216.7 441982.1 467551.7 495154.1 506331.1 316989.5
#> [25] 797387.2 512904.7 331589.7 329722.8 289721.7 321897.1 271584.2 473500.1
#> [33] 465522.8 477163.3 470845.3 895748.9 339310.8 847949.7 300447.5 427044.9
#> [41] 333256.7 333375.6 494308.3 440080.5 684948.3 418076.1
#> 
#> [[9]]
#>  [1]  865530.0  475190.4  388120.4  747957.6  452991.0 1082013.3  322755.3
#>  [8]  630840.9  303912.6  324339.3  949803.3  454697.6  315264.7  609536.7
#> [15]  784973.3  725955.0  675018.7  408531.6  576711.9  573709.3  610236.7
#> [22]  633510.9  609701.7  321248.7  906166.7  626878.9  347731.9  319013.3
#> [29]  309911.4  353840.4  305777.6  618594.0  576518.3  479248.3  445026.3
#> [36] 1019395.4  343840.9  960300.0  345164.3  543054.3  312503.3  352971.6
#> [43]  609536.7  556074.3  841490.0  441966.0
#> 
#> [[10]]
#>  [1] 910517.4 379381.9 345095.7 573058.9 445931.0 871640.3 275432.4 524742.3
#>  [9] 334764.1 334620.0 847131.0 375175.5 310964.8 481591.1 586742.1 636584.2
#> [17] 489518.3 388302.8 484926.3 491904.9 504044.6 510808.0 492998.6 313272.3
#> [25] 840081.5 521056.9 318278.0 301678.9 309618.5 348663.3 299189.7 513455.8
#> [33] 481636.8 468773.6 443040.4 860840.5 397635.5 885847.1 332463.9 480917.4
#> [41] 303973.1 337000.7 481591.1 459906.9 678961.9 421253.2
#> 
#> [[11]]
#>  [1]  912608.6  409764.0  389674.1  594533.3  465594.5 1027360.1  332949.7
#>  [8]  538012.8  370200.3  366945.7  872138.4  399415.9  327966.1  503731.3
#> [15]  617683.7  612400.4  508530.3  405478.7  490692.5  518354.2  526136.6
#> [22]  527566.2  515362.1  324936.5  786710.5  531214.9  344018.9  360283.0
#> [29]  334911.4  375113.3  318437.5  539243.9  499693.3  486128.9  480871.4
#> [36]  916988.6  383182.2  826161.0  352445.3  494533.4  363237.3  374152.5
#> [43]  503731.3  445929.8  681495.1  462471.2
#> 
#> [[12]]
#>  [1] 820968.33 146479.22  89027.52 487154.38 204771.21 605678.65  49402.14
#>  [8] 290857.54  65122.02  99482.41 837512.44 118466.72 121946.74 342358.52
#> [15] 495282.56 239708.64 390192.15  40516.22 281312.75 257112.28 277820.43
#> [22] 261655.03 341777.54 162805.91 489513.11 417596.35 181793.55  48613.06
#> [29] 230276.98 146271.93  65122.02 220844.70 211318.66 211388.85 209291.28
#> [36] 854237.37 239559.76 522928.35 232674.94 201468.20  50830.83  52712.37
#> [43] 342358.52 377113.90 445652.72 177343.04
#> 
#> [[13]]
#>  [1]  929327.0  426730.7  356176.8  552546.5  422908.1 1043625.4  299946.6
#>  [8]  522240.0  335788.9  333367.3  808490.9  383908.6  326419.6  466078.0
#> [15]  572248.6  574843.0  481616.2  367595.5  448906.7  465101.7  499324.6
#> [22]  519367.5  475139.3  316155.3  762004.6  513027.5  324136.9  324970.3
#> [29]  329188.1  353332.5  304547.7  516019.7  495080.8  489208.9  448101.0
#> [36]  848339.8  407377.7  788457.1  344106.9  447540.0  334692.1  353451.5
#> [43]  466078.0  424380.3  754514.5  406347.4
#> 
#> [[14]]
#>  [1]  783676.9  306688.8  226826.5  653550.5  325644.4  679645.9  204604.8
#>  [8]  401106.7  187012.9  216834.5 1363407.2  283057.9  219653.0  415705.7
#> [15]  678295.1  670621.5  447644.4  298618.5  425990.6  402966.9  427924.9
#> [22]  408444.5  506247.9  219044.3  566341.1  412059.1  240140.4  218996.5
#> [29]  207673.6  280051.9  158943.4  433999.7  427838.6  425767.7  425767.7
#> [36] 1350620.6  266208.5  680133.6  337626.5  362285.8  289170.2  160929.1
#> [43]  415705.7  424967.9  466916.3  310064.2
#> 
#> [[15]]
#>  [1]  665472.1  289635.1  253750.4  702953.1  252426.9  837321.9  158098.1
#>  [8]  448801.9  174547.1  154244.0 1123885.0  306975.5  248934.9  436793.1
#> [15]  706394.9  827258.3  600439.0  262641.6  377709.7  394150.5  419143.7
#> [22]  403053.9  554843.3  233229.2  724433.0  451378.9  364285.9  198649.3
#> [29]  255307.3  289421.4  206645.6  382939.0  365459.7  359835.5  359835.5
#> [36]  977060.7  325937.1  643479.7  267729.5  283812.3  171144.4  306583.6
#> [43]  436793.1  362324.3  382390.5  269881.3
#> 
#> [[16]]
#>  [1] 918834.8 442449.6 428782.9 621311.3 480258.4 863784.1 374815.5 548514.0
#>  [9] 397679.0 392768.2 860219.3 424830.2 385762.7 526160.6 633668.7 628563.4
#> [17] 555617.9 437300.7 518216.6 533536.5 518342.0 544237.7 538031.0 381169.5
#> [25] 853002.4 545789.2 412763.1 403040.1 400868.1 414855.1 384946.5 520581.4
#> [33] 521367.2 526123.0 492853.5 899968.6 454211.4 877723.0 410061.8 492918.6
#> [41] 400981.8 388554.8 526160.6 533191.0 671415.7 522481.3
#> 
#> [[17]]
#>  [1] 877507.9 284485.6 309926.0 945900.1 362064.5 668409.2 218880.0 577402.2
#>  [9] 256263.5 251840.0 971996.2 267246.4 286688.1 455527.9 945900.1 662043.8
#> [17] 552924.5 290309.1 375338.5 389964.8 398815.4 493477.8 516886.5 286688.1
#> [25] 602120.9 540363.3 321013.1 229898.1 327113.5 284485.6 256263.5 403990.0
#> [33] 378015.4 378015.4 352785.0 953859.7 289498.0 911957.3 288168.4 353422.5
#> [41] 242560.5 241848.9 455527.9 375338.5 703081.0 324377.6
#> 
#> [[18]]
#>  [1] 877507.9 284485.6 309926.0 945900.1 362064.5 668409.2 218880.0 577402.2
#>  [9] 256263.5 251840.0 971996.2 267246.4 286688.1 455527.9 945900.1 662043.8
#> [17] 552924.5 290309.1 375338.5 389964.8 398815.4 493477.8 516886.5 286688.1
#> [25] 602120.9 540363.3 321013.1 229898.1 327113.5 284485.6 256263.5 403990.0
#> [33] 378015.4 378015.4 352785.0 953859.7 289498.0 911957.3 288168.4 353422.5
#> [41] 242560.5 241848.9 455527.9 375338.5 703081.0 324377.6
#> 
#> [[19]]
#>  [1] 856915.4 266330.4 307704.2 920108.3 351283.9 618916.2 208330.8 581658.7
#>  [9] 243749.9 266428.0 939536.9 266330.4 287619.6 437478.7 920108.3 676308.0
#> [17] 550578.2 286167.0 376006.5 359640.9 391882.0 502265.5 475657.1 287619.6
#> [25] 600187.7 481987.6 337618.0 244688.5 365785.2 268075.4 243749.9 413902.0
#> [33] 385208.7 394636.2 354883.6 902054.1 293467.0 882116.7 289844.6 338432.5
#> [41] 238304.1 247196.8 437478.7 382327.3 723805.1 304507.0
#> 
#> [[20]]
#>  [1] 856915.4 266330.4 307704.2 920108.3 351283.9 618916.2 208330.8 581658.7
#>  [9] 243749.9 266428.0 939536.9 266330.4 287619.6 437478.7 920108.3 676308.0
#> [17] 550578.2 286167.0 376006.5 359640.9 391882.0 502265.5 475657.1 287619.6
#> [25] 600187.7 481987.6 337618.0 244688.5 365785.2 268075.4 243749.9 413902.0
#> [33] 385208.7 394636.2 354883.6 902054.1 293467.0 882116.7 289844.6 338432.5
#> [41] 238304.1 247196.8 437478.7 382327.3 723805.1 304507.0
#> 
#> [[21]]
#>  [1]  897514.6  283736.8  272127.7  921523.4  345044.2  652307.9  141502.9
#>  [8]  453432.2  197132.2  272231.5 1023169.4  283736.8  236665.3  425839.7
#> [15]  921523.4  648266.5  501597.8  225451.5  391738.8  298369.5  402356.3
#> [22]  412057.7  467412.9  236665.3  580505.1  435168.3  284556.5  230918.0
#> [29]  316127.9  283736.8  207218.1  426140.2  388914.4  412747.2  359709.6
#> [36]  957915.0  226624.9  895412.3  260466.9  330809.0  239939.6  238572.2
#> [43]  425839.7  390415.5  725291.8  242750.5
#> 
#> [[22]]
#>  [1] 799816.5 286165.9 310362.1 822614.9 396489.0 679567.4 293404.4 583206.3
#>  [9] 298762.6 310362.1 851312.1 286165.9 311374.7 471297.7 822614.9 729286.8
#> [17] 494746.2 372225.6 394167.9 418705.3 409368.6 548567.8 512870.8 311374.7
#> [25] 630001.7 511529.9 317389.1 294579.4 311374.7 286165.9 298762.6 472783.9
#> [33] 435558.1 435558.1 418607.8 851312.1 328673.7 822614.9 329499.3 446594.4
#> [41] 307345.8 310095.2 471297.7 397427.9 655665.4 389524.5
#> 
#> [[23]]
#>  [1] 978554.2 435820.0 414204.7 573965.3 397262.3 709335.7 293853.7 468943.7
#>  [9] 320383.7 340124.1 790913.9 432964.9 320295.5 454479.0 572426.6 526052.9
#> [17] 487584.4 375498.3 489953.1 430329.4 543466.7 442812.3 518088.6 301263.3
#> [25] 568724.1 466450.0 297995.4 326510.6 302221.1 414774.2 346315.1 436456.1
#> [33] 435629.0 440530.0 440530.0 828632.8 358234.2 610032.8 385565.7 410739.8
#> [41] 325412.2 337256.1 454479.0 496628.4 547198.0 398502.1
#> 
#> [[24]]
#>  [1]  875060.0  313115.0  268198.1  832465.2  441433.2  684044.5  121951.3
#>  [8]  308331.2  227824.0  335177.8 1201193.8  313115.0  157178.2  501854.3
#> [15]  832465.2  617407.3  442459.9  205899.9  421116.9  403094.0  428519.0
#> [22]  273942.7  543427.4  157178.2  699063.9  430205.2  269261.5  201832.6
#> [29]  236640.8  312571.4  140829.8  402104.8  364879.0  355628.6  318234.0
#> [36] 1025408.5  214953.3  843937.6  213946.1  298169.3  211582.8  277628.7
#> [43]  501854.3  420503.4  709642.6  229552.9
#> 
#> [[25]]
#>  [1]  858595.9  327234.6  316795.4  898217.4  405122.8  696424.2  152146.3
#>  [8]  361772.9  222254.0  319248.9 1162711.6  309995.4  172333.6  471142.5
#> [15]  898217.4  639472.1  501622.7  223575.4  413009.7  393563.0  441255.2
#> [22]  301184.4  527997.0  172333.6  666708.1  451556.1  293808.1  241620.9
#> [29]  254855.3  328262.7  179238.3  404202.4  378227.7  358797.4  342636.0
#> [36] 1023990.3  204187.2  872999.8  184434.8  292473.8  230639.8  246329.3
#> [43]  471142.5  419788.0  703590.4  257643.9
#> 
#> [[26]]
#>  [1]  673844.2  210551.3  168234.9  694597.2  295514.4  863177.3  160198.1
#>  [8]  342995.9  199223.5  229471.8 1761327.9  248779.0  318522.9  312004.8
#> [15]  718109.5  622068.2  526529.9  203730.9  497046.0  371340.8  507015.5
#> [22]  375452.4  475991.3  290572.5  550236.2  383555.1  249263.5  269572.5
#> [29]  332173.8  148782.9  154485.8  419502.9  433594.6  454324.8  454324.8
#> [36]  819360.5  406019.3  587408.4  296158.7  305926.8  270527.4  174489.6
#> [43]  312004.8  442224.0  495536.9  253461.4
#> 
#> [[27]]
#>  [1] 869135.6 485080.6 453294.5 496790.9 461372.7 984729.5 422456.3 516738.2
#>  [9] 437978.2 447866.9 813155.1 477315.8 424745.5 476038.9 537312.1 590011.5
#> [17] 465176.5 432855.4 501052.3 496101.4 530730.6 516738.2 478080.4 421160.5
#> [25] 697380.7 490789.2 430972.8 428477.8 415549.9 464360.2 415575.1 528207.1
#> [33] 492409.2 471688.8 471688.8 813155.1 475701.1 749539.9 428453.8 496807.9
#> [41] 432889.6 458371.8 476038.9 471623.0 612905.6 471872.2
#> 
#> [[28]]
#>  [1]  790058.0  472762.3  456181.5  513215.7  465724.9 1095857.6  443462.0
#>  [8]  511030.8  450816.8  443815.8  791702.8  469624.4  444144.7  481809.0
#> [15]  559868.5  572491.4  479967.6  454125.9  489679.8  489738.3  528686.6
#> [22]  511030.8  487846.1  444144.7  702053.6  488871.2  452159.5  449999.3
#> [29]  444242.6  457468.3  436637.2  518029.9  488692.2  475134.1  475134.1
#> [36]  788537.5  463643.8  711181.1  453021.8  484049.4  449999.3  461899.2
#> [43]  481809.0  521238.9  621468.4  508957.5
#> 
#> [[29]]
#>  [1] 725074.8 388269.2 405325.5 689577.1 480020.7 699051.9 383244.0 507876.5
#>  [9] 388269.2 405325.5 704070.6 388269.2 386820.0 527236.1 689577.1 704585.5
#> [17] 545938.5 452235.9 485562.2 496862.0 496862.0 507876.5 550963.8 386820.0
#> [25] 678562.6 555279.2 386820.0 388269.2 386820.0 388269.2 388269.2 506588.7
#> [33] 485562.2 485562.2 485562.2 704070.6 403145.0 689577.1 391845.2 517888.5
#> [41] 388269.2 400300.2 527236.1 485562.2 598298.4 468560.9
#> 
#> [[30]]
#>  [1] 725074.8 388269.2 405325.5 689577.1 480020.7 699051.9 383244.0 507876.5
#>  [9] 388269.2 405325.5 704070.6 388269.2 386820.0 527236.1 689577.1 704585.5
#> [17] 545938.5 452235.9 485562.2 496862.0 496862.0 507876.5 550963.8 386820.0
#> [25] 678562.6 555279.2 386820.0 388269.2 386820.0 388269.2 388269.2 506588.7
#> [33] 485562.2 485562.2 485562.2 704070.6 403145.0 689577.1 391845.2 517888.5
#> [41] 388269.2 400300.2 527236.1 485562.2 598298.4 468560.9
#> 
#> [[31]]
#>  [1] 634638.6 145341.9 184245.6 293018.7 210117.8 634638.6 142841.9 406531.2
#>  [9] 184245.6 184245.6 634638.6 184245.6 184245.6 406531.2 406531.2 257836.8
#> [17] 406531.2 343755.6 134258.9 210117.8 343755.6 406531.2 406531.2 184245.6
#> [25] 634638.6 406531.2 184245.6 184245.6 184245.6 184245.6 184245.6 343755.6
#> [33] 343755.6 343755.6 343755.6 634638.6 184245.6 634638.6 184245.6 278411.6
#> [41] 184245.6 184245.6 406531.2 134258.9 293018.7 343755.6
#> 
#> [[32]]
#>  [1] 634638.6 145341.9 184245.6 293018.7 210117.8 634638.6 142841.9 406531.2
#>  [9] 184245.6 184245.6 634638.6 184245.6 184245.6 406531.2 406531.2 257836.8
#> [17] 406531.2 343755.6 134258.9 210117.8 343755.6 406531.2 406531.2 184245.6
#> [25] 634638.6 406531.2 184245.6 184245.6 184245.6 184245.6 184245.6 343755.6
#> [33] 343755.6 343755.6 343755.6 634638.6 184245.6 634638.6 184245.6 278411.6
#> [41] 184245.6 184245.6 406531.2 134258.9 293018.7 343755.6
#> 
#> [[33]]
#>  [1] 634638.6 181212.8 181212.8 293018.7 210117.8 634638.6 142841.9 397631.8
#>  [9] 181212.8 181212.8 634638.6 181212.8 181212.8 397631.8 397631.8 248937.4
#> [17] 397631.8 343755.6 134258.9 210117.8 343755.6 397631.8 397631.8 181212.8
#> [25] 634638.6 397631.8 181212.8 181212.8 181212.8 181212.8 181212.8 343755.6
#> [33] 343755.6 343755.6 343755.6 634638.6 181212.8 634638.6 181212.8 278411.6
#> [41] 181212.8 181212.8 397631.8 134258.9 293018.7 343755.6
#> 
#> [[34]]
#>  [1] 614963.9 207537.9 207537.9 318733.2 318733.2 614963.9 207537.9 318733.2
#>  [9] 207537.9 207537.9 614963.9 207537.9 207537.9 318733.2 318733.2 318733.2
#> [17] 318733.2 318733.2 318733.2 318733.2 318733.2 318733.2 318733.2 207537.9
#> [25] 614963.9 318733.2 207537.9 207537.9 207537.9 207537.9 207537.9 318733.2
#> [33] 318733.2 318733.2 318733.2 614963.9 207537.9 614963.9 207537.9 318733.2
#> [41] 207537.9 207537.9 318733.2 318733.2 318733.2 318733.2
#> 
#> [[35]]
#>  [1] 785583.3 278263.4 278263.4 471314.3 278263.4 785583.3 278263.4 471314.3
#>  [9] 278263.4 278263.4 785583.3 278263.4 278263.4 278263.4 471314.3 471314.3
#> [17] 278263.4 278263.4 278263.4 471314.3 471314.3 471314.3 278263.4 278263.4
#> [25] 785583.3 471314.3 278263.4 278263.4 278263.4 278263.4 278263.4 278263.4
#> [33] 278263.4 278263.4 278263.4 785583.3 471314.3 785583.3 278263.4 471314.3
#> [41] 278263.4 278263.4 278263.4 278263.4 471314.3 471314.3
#> 
#> [[36]]
#>  [1] 785583.3 278263.4 278263.4 471314.3 278263.4 785583.3 278263.4 471314.3
#>  [9] 278263.4 278263.4 785583.3 278263.4 278263.4 278263.4 471314.3 471314.3
#> [17] 278263.4 278263.4 278263.4 471314.3 471314.3 471314.3 278263.4 278263.4
#> [25] 785583.3 471314.3 278263.4 278263.4 278263.4 278263.4 278263.4 278263.4
#> [33] 278263.4 278263.4 278263.4 785583.3 471314.3 785583.3 278263.4 471314.3
#> [41] 278263.4 278263.4 278263.4 278263.4 471314.3 471314.3
#> 
#> [[37]]
#>  [1] 785583.3 278263.4 278263.4 471314.3 278263.4 785583.3 278263.4 471314.3
#>  [9] 278263.4 278263.4 785583.3 278263.4 278263.4 278263.4 471314.3 471314.3
#> [17] 278263.4 278263.4 278263.4 471314.3 471314.3 471314.3 278263.4 278263.4
#> [25] 785583.3 471314.3 278263.4 278263.4 278263.4 278263.4 278263.4 278263.4
#> [33] 278263.4 278263.4 278263.4 785583.3 471314.3 785583.3 278263.4 471314.3
#> [41] 278263.4 278263.4 278263.4 278263.4 471314.3 471314.3
#> 
#> [[38]]
#>  [1] 785583.3 278263.4 278263.4 471314.3 278263.4 785583.3 278263.4 471314.3
#>  [9] 278263.4 278263.4 785583.3 278263.4 278263.4 278263.4 471314.3 471314.3
#> [17] 278263.4 278263.4 278263.4 471314.3 471314.3 471314.3 278263.4 278263.4
#> [25] 785583.3 471314.3 278263.4 278263.4 278263.4 278263.4 278263.4 278263.4
#> [33] 278263.4 278263.4 278263.4 785583.3 471314.3 785583.3 278263.4 471314.3
#> [41] 278263.4 278263.4 278263.4 278263.4 471314.3 471314.3
#> 
#> [[39]]
#>  [1] 785583.3 278263.4 278263.4 471314.3 278263.4 785583.3 278263.4 471314.3
#>  [9] 278263.4 278263.4 785583.3 278263.4 278263.4 278263.4 471314.3 471314.3
#> [17] 278263.4 278263.4 278263.4 471314.3 471314.3 471314.3 278263.4 278263.4
#> [25] 785583.3 471314.3 278263.4 278263.4 278263.4 278263.4 278263.4 278263.4
#> [33] 278263.4 278263.4 278263.4 785583.3 471314.3 785583.3 278263.4 471314.3
#> [41] 278263.4 278263.4 278263.4 278263.4 471314.3 471314.3
#> 
#> [[40]]
#>  [1] 785583.3 278263.4 278263.4 471314.3 278263.4 785583.3 278263.4 471314.3
#>  [9] 278263.4 278263.4 785583.3 278263.4 278263.4 278263.4 471314.3 471314.3
#> [17] 278263.4 278263.4 278263.4 471314.3 471314.3 471314.3 278263.4 278263.4
#> [25] 785583.3 471314.3 278263.4 278263.4 278263.4 278263.4 278263.4 278263.4
#> [33] 278263.4 278263.4 278263.4 785583.3 471314.3 785583.3 278263.4 471314.3
#> [41] 278263.4 278263.4 278263.4 278263.4 471314.3 471314.3
#> 
#> [[41]]
#>  [1] 785583.3 278263.4 278263.4 471314.3 278263.4 785583.3 278263.4 471314.3
#>  [9] 278263.4 278263.4 785583.3 278263.4 278263.4 278263.4 471314.3 471314.3
#> [17] 278263.4 278263.4 278263.4 471314.3 471314.3 471314.3 278263.4 278263.4
#> [25] 785583.3 471314.3 278263.4 278263.4 278263.4 278263.4 278263.4 278263.4
#> [33] 278263.4 278263.4 278263.4 785583.3 471314.3 785583.3 278263.4 471314.3
#> [41] 278263.4 278263.4 278263.4 278263.4 471314.3 471314.3
#> 
#> [[42]]
#>  [1] 785583.3 278263.4 278263.4 471314.3 278263.4 785583.3 278263.4 471314.3
#>  [9] 278263.4 278263.4 785583.3 278263.4 278263.4 278263.4 471314.3 471314.3
#> [17] 278263.4 278263.4 278263.4 471314.3 471314.3 471314.3 278263.4 278263.4
#> [25] 785583.3 471314.3 278263.4 278263.4 278263.4 278263.4 278263.4 278263.4
#> [33] 278263.4 278263.4 278263.4 785583.3 471314.3 785583.3 278263.4 471314.3
#> [41] 278263.4 278263.4 278263.4 278263.4 471314.3 471314.3
#> 
#> [[43]]
#>  [1] 785583.3 278263.4 278263.4 471314.3 278263.4 785583.3 278263.4 471314.3
#>  [9] 278263.4 278263.4 785583.3 278263.4 278263.4 278263.4 471314.3 471314.3
#> [17] 278263.4 278263.4 278263.4 471314.3 471314.3 471314.3 278263.4 278263.4
#> [25] 785583.3 471314.3 278263.4 278263.4 278263.4 278263.4 278263.4 278263.4
#> [33] 278263.4 278263.4 278263.4 785583.3 471314.3 785583.3 278263.4 471314.3
#> [41] 278263.4 278263.4 278263.4 278263.4 471314.3 471314.3
#> 
#> [[44]]
#>  [1] 785583.3 278263.4 278263.4 471314.3 278263.4 785583.3 278263.4 471314.3
#>  [9] 278263.4 278263.4 785583.3 278263.4 278263.4 278263.4 471314.3 471314.3
#> [17] 278263.4 278263.4 278263.4 471314.3 471314.3 471314.3 278263.4 278263.4
#> [25] 785583.3 471314.3 278263.4 278263.4 278263.4 278263.4 278263.4 278263.4
#> [33] 278263.4 278263.4 278263.4 785583.3 471314.3 785583.3 278263.4 471314.3
#> [41] 278263.4 278263.4 278263.4 278263.4 471314.3 471314.3
#> 
#> [[45]]
#>  [1] 1176294.6  305065.1  379339.0  880370.9  315576.7  411699.4  220734.5
#>  [8]  997415.4  315576.7  237815.2  775884.1  305065.1  263112.5  316886.1
#> [15]  880370.9  769004.7  334604.1  220734.5  305065.1  359986.6  523946.3
#> [22]  934181.8  316886.1  263112.5  523946.3  502648.9  263112.5  237815.2
#> [29]  263112.5  315576.7  237815.2  592665.7  592665.7  570820.0  570820.0
#> [36]  774094.4  465758.6  974101.2  314280.0  592833.2  220734.5  246353.3
#> [43]  316886.1  304191.2 1495153.0  419893.2
#> 
#> [[46]]
#>  [1] 679153.5 438514.2 435077.2 683300.9 415670.2 560986.7 358649.3 788861.0
#>  [9] 415670.2 385414.0 592770.5 438514.2 369995.3 422947.3 683300.9 701639.0
#> [17] 400438.9 358649.3 438514.2 470172.8 567403.7 637754.0 422947.3 369995.3
#> [25] 567403.7 494237.9 369995.3 385414.0 369995.3 415670.2 385414.0 577805.6
#> [33] 577805.6 587879.4 587879.4 610230.8 454629.5 636342.5 416306.1 609251.4
#> [41] 358649.3 366636.9 422947.3 404813.4 749254.3 447168.3
#> 
#> [[47]]
#>  [1]  794362.2  434567.4  441869.5  870674.6  403637.7  450304.2  238617.3
#>  [8] 1095999.7  403637.7  309266.5  713190.4  434567.4  259761.9  353781.7
#> [15]  870674.6  885927.7  301559.9  238617.3  434567.4  434370.7  497209.1
#> [22]  961902.3  353781.7  259761.9  497209.1  575465.1  259761.9  309266.5
#> [29]  259761.9  403637.7  309266.5  670986.2  670986.2  654262.5  654262.5
#> [36]  716420.8  435819.8  706214.4  373170.2  623554.4  238617.3  280625.8
#> [43]  353781.7  359545.7 1388443.3  411371.1
#> 
#> [[48]]
#>  [1]  705450.2  456292.1  474992.0 1258461.1  399233.3  425766.9  303620.6
#>  [8] 1298996.2  399233.3  329696.8  744060.8  456292.1  307570.0  426622.3
#> [15] 1258461.1  941209.1  435561.0  303620.6  456292.1  414126.1  492690.9
#> [22] 1027765.7  426622.3  307570.0  492690.9  620000.0  307570.0  329696.8
#> [29]  307570.0  399233.3  329696.8  754941.8  754941.8  686850.3  686850.3
#> [36]  982653.0  469681.9  904498.2  535718.9  666642.1  303620.6  309346.9
#> [43]  426622.3  386350.6 1211442.4  500089.2
#> 
#> [[49]]
#>  [1]  598394.3  486975.0  464710.7  730463.1  431406.7  481890.5  288790.6
#>  [8] 1131596.9  431406.7  388736.2  745766.8  486975.0  307225.9  443677.6
#> [15]  730463.1 1171997.0  426619.3  288790.6  486975.0  465733.1  450649.2
#> [22] 1209473.2  443677.6  307225.9  450649.2  614816.0  307225.9  388736.2
#> [29]  307225.9  431406.7  388736.2  468634.7  468634.7  522329.1  522329.1
#> [36]  741672.0  452879.8  479414.0  406734.3  547818.9  288790.6  317168.6
#> [43]  443677.6  424933.3 1325541.2  405322.7
#> 
#> [[50]]
#>  [1] 393971.4 145736.0 145736.0 212453.2 212453.2 393971.4 145736.0 212453.2
#>  [9] 145736.0 145736.0 393971.4 145736.0 145736.0 212453.2 212453.2 212453.2
#> [17] 212453.2 212453.2 212453.2 212453.2 212453.2 212453.2 212453.2 145736.0
#> [25] 393971.4 212453.2 145736.0 145736.0 145736.0 145736.0 145736.0 212453.2
#> [33] 212453.2 212453.2 212453.2 393971.4 145736.0 393971.4 145736.0 212453.2
#> [41] 145736.0 145736.0 212453.2 212453.2 212453.2 212453.2
#> 
#> [[51]]
#>  [1] 395125.38 119789.91 119789.91 300703.81 147684.41 395125.38  89867.15
#>  [8] 227867.06 119789.91 119789.91 395125.38 119789.91 119789.91 177734.48
#> [15] 300703.81 211487.19 177734.48 227867.06 141375.41 147684.41 227867.06
#> [22] 227867.06 177734.48 119789.91 395125.38 177734.48 119789.91 119789.91
#> [29] 119789.91 119789.91 119789.91 227867.06 227867.06 227867.06 227867.06
#> [36] 395125.38 119789.91 395125.38 119789.91 170984.47 119789.91 119789.91
#> [43] 177734.48 141375.41 300703.81 227867.06
#> 
#> [[52]]
#>  [1] 395125.38 119789.91 119789.91 300703.81 147684.41 395125.38  89867.15
#>  [8] 227867.06 119789.91 119789.91 395125.38 119789.91 119789.91 177734.48
#> [15] 300703.81 211487.19 177734.48 227867.06 141375.41 147684.41 227867.06
#> [22] 227867.06 177734.48 119789.91 395125.38 177734.48 119789.91 119789.91
#> [29] 119789.91 119789.91 119789.91 227867.06 227867.06 227867.06 227867.06
#> [36] 395125.38 119789.91 395125.38 119789.91 170984.47 119789.91 119789.91
#> [43] 177734.48 141375.41 300703.81 227867.06
#> 
#> [[53]]
#>  [1] 153044.05  51612.42  51612.42  93601.97  66874.42 153044.05  51612.42
#>  [8]  93601.97  51612.42  51612.42 153044.05  51612.42  39402.02  93601.97
#> [15]  93601.97  63863.09  93601.97  93601.97  64771.42  66874.42  93601.97
#> [22]  93601.97  93601.97  39402.02 153044.05  93601.97  51612.42  51612.42
#> [29]  39402.02  51612.42  51612.42  93601.97  93601.97  93601.97  93601.97
#> [36] 153044.05  51612.42 153044.05  51612.42  93601.97  51612.42  51612.42
#> [43]  93601.97  64771.42  93601.97  93601.97
#> 
#> [[54]]
#>  [1] 151951.72  55650.04  55650.04  77889.11  77889.11 151951.72  55650.04
#>  [8]  77889.11  55650.04  55650.04 151951.72  55650.04  55650.04  77889.11
#> [15]  77889.11  77889.11  77889.11  77889.11  77889.11  77889.11  77889.11
#> [22]  77889.11  77889.11  55650.04 151951.72  77889.11  55650.04  55650.04
#> [29]  55650.04  55650.04  55650.04  77889.11  77889.11  77889.11  77889.11
#> [36] 151951.72  55650.04 151951.72  55650.04  77889.11  55650.04  55650.04
#> [43]  77889.11  77889.11  77889.11  77889.11
#> 
#> [[55]]
#>  [1] 151951.72  55650.04  55650.04  77889.11  77889.11 151951.72  55650.04
#>  [8]  77889.11  55650.04  55650.04 151951.72  55650.04  55650.04  77889.11
#> [15]  77889.11  77889.11  77889.11  77889.11  77889.11  77889.11  77889.11
#> [22]  77889.11  77889.11  55650.04 151951.72  77889.11  55650.04  55650.04
#> [29]  55650.04  55650.04  55650.04  77889.11  77889.11  77889.11  77889.11
#> [36] 151951.72  55650.04 151951.72  55650.04  77889.11  55650.04  55650.04
#> [43]  77889.11  77889.11  77889.11  77889.11
#> 
#> [[56]]
#>  [1]  908628.5  353319.2  353576.0  598448.3  416043.4 1053125.9  297593.0
#>  [8]  501173.6  312963.6  332148.7  977524.7  351906.7  292329.9  490196.6
#> [15]  636441.0  566192.0  498778.5  372076.3  447616.9  450271.3  467830.6
#> [22]  494312.3  511360.9  290173.5  776784.4  529508.7  332634.3  322346.2
#> [29]  299750.3  333485.8  299278.3  498149.3  467084.3  468720.0  433014.2
#> [36]  987535.8  396088.4  830312.6  342644.6  432926.2  328061.1  337369.0
#> [43]  490196.6  438744.4  698721.3  404989.2
#> 
#> [[57]]
#>  [1] 841733.1 272024.8 196139.1 649778.0 259987.7 602342.6 158898.9 334222.6
#>  [9] 132562.4 156799.5 908102.9 182203.3 185133.1 486959.2 808150.4 368633.3
#> [17] 429946.5 290228.7 426903.4 238105.9 413348.7 420629.8 595553.8 185133.1
#> [25] 628360.7 512496.3 261555.8 231354.6 192982.6 245285.8 134226.9 362937.8
#> [33] 396700.7 423545.1 339828.5 913971.4 299952.3 738324.0 298139.4 346001.8
#> [41] 177943.5 198277.4 486959.2 472360.4 685895.7 404003.9
#> 
#> [[58]]
#>  [1] 785583.3 278263.4 278263.4 471314.3 278263.4 785583.3 278263.4 471314.3
#>  [9] 278263.4 278263.4 785583.3 278263.4 278263.4 278263.4 471314.3 471314.3
#> [17] 278263.4 278263.4 278263.4 471314.3 471314.3 471314.3 278263.4 278263.4
#> [25] 785583.3 471314.3 278263.4 278263.4 278263.4 278263.4 278263.4 278263.4
#> [33] 278263.4 278263.4 278263.4 785583.3 471314.3 785583.3 278263.4 471314.3
#> [41] 278263.4 278263.4 278263.4 278263.4 471314.3 471314.3
#> 
#> [[59]]
#>  [1]  959173.1  313863.2  270918.2  848925.3  413044.7  725538.7  154212.9
#>  [8]  436472.3  187904.6  265765.2 1151605.0  313863.2  149477.8  514855.2
#> [15]  848925.3  599140.0  453480.4  213594.0  421349.0  339679.9  442439.2
#> [22]  375594.0  495505.0  149477.8  633675.6  441460.2  276301.5  268228.7
#> [29]  309644.0  313863.2  130692.0  441455.4  376270.0  376270.0  344264.3
#> [36]  947404.9  256276.8  853204.7  201796.9  300502.3  210537.5  268577.2
#> [43]  514855.2  393244.3  700066.7  251865.0
#> 
#> [[60]]
#>  [1]  964142.1  451228.0  363204.2  616986.9  370035.6 1143761.7  350726.6
#>  [8]  494783.4  365138.9  357595.1 1972688.1  399215.5  240057.3  506949.5
#> [15]  766667.0  484067.4  691149.6  396450.1  468255.3  409239.7  527426.1
#> [22]  410931.9  560146.2  222562.7  687046.0  440153.0  252139.6  304019.2
#> [29]  168579.3  456134.3  328721.6  575885.5  450193.0  475547.3  450644.2
#> [36] 1151539.8  252335.9  552756.7  329444.6  393930.1  350675.2  341014.3
#> [43]  506949.5  465892.2 1081793.2  424569.7
#>