Train models with forester

The `train()` is the core function of this package. The only obligatory arguments are `data` and `target`. Setting and changing other arguments will affect model validation strategy, tested model families and so on.

train(
  data,
  y,
  type = "guess",
  engine = c("ranger", "xgboost", "decision_tree", "lightgbm", "catboost"),
  loss = "default",
  validation = "default",
  tuning = "default",
  keep = FALSE
)

Arguments

data
`data.frame` or `matrix` - data which will be used to build models. By default model will be trained on all columns in the `data`

y
target variable. It can be either (1) a vector of the same number of observations as `data` or (2) a character name of variable in the `data` that contains the target variable

type
a character, one of `classification`/`regression`/`guess`. sets the type of the task. If `guess` (the default option) then forester will figure out `type` based on the number of unique values in the `y` variable

engine
a vector of tree-based models that shall be testes. Possible values are: `ranger`, `xgboost`, `decision_tree`, `lightgbm`, `catboost`. All models from this vector will be trained and the best one will be returned.

loss
a character with the name of the loss function. If `default` then the default loss function for a given task will be used

validation
validation/model selection a character of data.frame. If data.frame then this data will be used for model assessment and selection. If character, then it will be a name of the hyperparameter.

tuning
a character with the name of the tuning procedure. one out of `default`/`portfolio`/`bayesian`/`random search`. If `default` then no tuning, just default hyperparameteres are used. If `portfolio` then static portfolio of hyperparametes are tried. If `bayesian` then Bayesian optimization is performed. If `random search` then random search is performed.

keep
shall all models be returned (`keep = TRUE`) or only the best one (`keep = FALSE`, default option)?

Value

Examples