Fertility dataset

fertility 100 volunteers provide a semen sample analyzed according to the WHO 2010 criteria. Sperm concentration are related to socio-demographic data, environmental factors, health status, and life habits

data(fertility)

Format

A data frame with 100 rows and 10 variables:

season

numeric, Season in which the analysis was performed. 1) winter, 2) spring, 3) Summer, 4) fall. (-1, -0.33, 0.33, 1)

age

numeric, Age at the time of analysis. 18-36 (0, 1)

childish_diseases

numeric, Childish diseases (ie , chicken pox, measles, mumps, polio) 1) yes, 2) no. (0, 1)

accident_or_serious_trauma

numeric,Accident or serious trauma 1) yes, 2) no. (0, 1)

surgerical_intervention

numeric, Surgical intervention 1) yes, 2) no. (0, 1)

high_fevers_last_year

numeric, High fevers in the last year 1) less than three months ago, 2) more than three months ago, 3) no. (-1, 0, 1)

alcohol_consumption

numeric,Frequency of alcohol consumption 1) several times a day, 2) every day, 3) several times a week, 4) once a week, 5) hardly ever or never (0, 1)

smoking

numeric, Smoking habit 1) never, 2) occasional 3) daily. (-1, 0, 1)

sitting_hours

numeric, Number of hours spent sitting per day ene-16 (0, 1)

diagnosis

factor, Diagnosis normal (1), altered (2)

Source

Data from UCI https://archive.ics.uci.edu/ml/datasets/Fertility