Perform predictions on new data

This function is used when the user created the models and evaluated them on a dataset x and later on, he got new observations and wants to predict the target value for them. The `predict_new()` function quickly normalizes their format, like a subset of columns that were deleted for the models and makes predictions. It is strongly advised that the new dataset shouldn't have any missing values.

predict_new(train_out, data, verbose = TRUE)

Arguments

train_out

The output of the `train()` function.

data

A dataset with the same format as the one used in `train()` function, but with new observations.

verbose

A logical value, if set to TRUE, provides all information about the process, if FALSE gives none.

Value

A list of predictions.

Examples

data(lisbon)
lisbon_train <- lisbon[1:200, ]
lisbon_new   <- lisbon[201:246, ]
out          <- train(lisbon_train, 'Price', verbose = FALSE)
preds        <- predict_new(out, lisbon_new)
preds
#> $ranger_model
#>  [1]  871253.0  427142.9  382587.8  709802.1  458485.4 1015209.3  345861.6
#>  [8]  536147.1  352022.6  377409.6  918831.4  419788.5  294052.3  516316.5
#> [15]  724808.8  744794.8  512967.4  433608.4  508261.5  495013.3  533399.5
#> [22]  535579.2  527416.5  295050.2  792472.5  535075.2  339495.3  343007.7
#> [29]  310420.3  375399.7  332928.2  538699.4  480719.5  476726.9  470624.1
#> [36]  969609.1  341982.2  876693.6  365636.4  546217.3  366473.9  376226.3
#> [43]  516417.3  459597.6  795025.3  443994.5
#> 
#> $xgboost_model
#>  [1] 622825.9 277683.8 318482.8 552897.0 335186.6 590938.8 263968.0 469506.6
#>  [9] 272111.2 294926.4 880792.2 258966.4 197011.1 498647.9 552897.0 451580.2
#> [17] 454743.7 398708.4 460387.1 332922.4 459798.3 469506.6 495002.9 215728.4
#> [25] 658395.8 468342.5 299552.6 265084.0 212588.1 277683.8 267146.4 368435.4
#> [33] 420256.3 420256.3 420256.3 882791.1 282124.8 827665.9 288110.1 508150.4
#> [41] 276577.5 320290.6 498647.9 429986.2 529892.1 396400.8
#> 
#> $decision_tree_model
#>  [1] 1049571.4  195018.2  195018.2 1049571.4  560642.9  722674.2  309000.0
#>  [8]  395181.8  195018.2  309000.0  722674.2  195018.2  309000.0  560642.9
#> [15] 1049571.4  722674.2  560642.9  395181.8  395181.8  395181.8  395181.8
#> [22]  560642.9  560642.9  195018.2  722674.2  560642.9  195018.2  195018.2
#> [29]  195018.2  195018.2  195018.2  395181.8  560642.9  395181.8  395181.8
#> [36]  912500.0  309000.0  722674.2  309000.0  395181.8  309000.0  195018.2
#> [43]  395181.8  395181.8 1049571.4  395181.8
#> 
#> $lightgbm_model
#>  [1]  761311.1  209011.3  207970.5  781287.7  307859.0  639103.7  141062.3
#>  [8]  627511.6  190961.5  221635.9  950389.1  189839.6  253381.4  408062.8
#> [15]  781287.7  410452.3  601289.6  229680.8  313442.5  335200.5  295204.0
#> [22]  608500.4  486414.6  292773.0  610240.2  525420.6  186528.6  161682.3
#> [29]  299220.2  209011.3  174701.3  358725.9  358725.9  358725.9  358725.9
#> [36] 1670111.3  218043.9  865279.7  317952.9  267193.5  238572.8  161829.2
#> [43]  408062.8  313442.5  507112.1  235102.7
#> 
#> $ranger_RS_1
#>  [1]  953115.4  393193.3  385911.7  695756.8  446100.1 1095257.8  356890.5
#>  [8]  538401.4  343758.9  365827.3  905751.4  401333.4  337398.5  492281.3
#> [15]  704406.0  732777.5  501729.8  419233.6  503133.9  459484.4  525878.5
#> [22]  527359.8  508460.5  329275.0  769356.4  513858.3  374132.2  352433.6
#> [29]  321506.3  375011.2  346280.9  514645.0  469788.7  459361.2  457505.7
#> [36]  951203.8  368730.3  853883.1  382876.5  545657.9  367464.9  371119.2
#> [43]  492281.3  506807.5  830235.3  489293.5
#> 
#> $ranger_RS_2
#>  [1]  897566.3  395400.4  362707.7  731759.9  444126.0 1116887.5  350062.1
#>  [8]  511173.8  328237.1  376007.9  859493.2  397868.0  318678.2  490428.6
#> [15]  731138.6  792013.3  497769.2  429435.4  466039.7  494994.2  499172.3
#> [22]  512722.7  493179.9  310770.8  773727.1  520060.5  357519.5  363821.9
#> [29]  303273.8  357317.3  323038.1  581758.4  480430.0  474841.9  476077.4
#> [36]  903108.2  385274.0  871412.8  342329.5  538645.1  376096.1  366892.6
#> [43]  490428.6  442930.6  828416.8  439961.2
#> 
#> $ranger_RS_3
#>  [1] 864085.2 413950.0 373226.4 708139.0 449984.2 945573.6 333638.6 515805.2
#>  [9] 346378.9 360804.9 874798.9 411899.3 327512.5 481037.3 734312.5 718132.5
#> [17] 500394.5 425408.9 487001.6 480926.4 523928.6 524006.4 483414.5 320813.7
#> [25] 812882.0 496273.8 338925.0 333580.6 331588.9 377593.9 343504.6 511393.6
#> [33] 469558.6 450207.8 449238.2 934849.3 349659.1 898172.8 351232.9 535058.7
#> [41] 341589.4 372962.4 481037.3 464888.9 745354.3 437089.8
#> 
#> $ranger_RS_4
#>  [1] 867043.4 522067.7 495008.0 654476.1 502045.4 790089.1 480942.0 536549.5
#>  [9] 480942.0 480942.0 743752.2 517696.1 462618.5 508537.4 654476.1 762387.6
#> [17] 499832.8 497945.6 543171.1 536549.5 543382.7 544507.8 508537.4 462618.5
#> [25] 739983.0 520242.1 476681.7 476681.7 462618.5 492146.8 487775.2 560398.4
#> [33] 531966.3 502045.4 502045.4 764933.6 475542.2 746161.6 481077.0 564981.6
#> [41] 480942.0 480942.0 508537.4 599919.8 774470.4 589198.4
#> 
#> $ranger_RS_5
#>  [1]  837674.0  368151.1  336052.1  687724.9  438857.2 1034615.5  327719.4
#>  [8]  499236.5  335693.8  364111.8  924623.6  367136.5  285037.3  489232.1
#> [15]  692931.2  672691.8  496602.2  417315.7  465431.3  472769.3  487162.3
#> [22]  519634.1  507964.2  281877.8  768843.1  502758.9  333607.5  320438.1
#> [29]  293066.2  332941.0  311056.2  521855.8  455333.5  454137.3  447152.2
#> [36]  948070.8  347055.0  850188.4  353382.8  514806.1  338788.7  355234.0
#> [43]  489232.1  442478.8  745360.9  455437.5
#> 
#> $ranger_RS_6
#>  [1] 893508.7 351148.1 347031.3 735682.4 400423.1 973197.5 299770.4 529544.6
#>  [9] 310045.5 321777.5 854930.7 350988.1 323386.5 429933.1 732233.0 755790.1
#> [17] 468732.5 373597.3 435437.8 419595.2 457121.1 533657.1 436226.6 284705.6
#> [25] 773744.7 460429.1 306461.6 304161.3 291470.5 338746.9 298337.8 516027.5
#> [33] 449145.9 416539.1 416539.1 884602.5 325884.3 911450.5 307211.5 519822.2
#> [41] 308088.8 341983.0 429933.1 442044.4 819964.9 404722.7
#> 
#> $ranger_RS_7
#>  [1]  912893.5  387608.5  357248.9  682710.4  420718.7 1123608.6  300894.1
#>  [8]  488119.3  306129.7  330629.0  946493.9  393710.0  268180.0  497538.8
#> [15]  688751.4  696859.0  513575.7  395650.3  481955.4  454725.3  497928.8
#> [22]  499238.9  526931.7  257021.9  787115.9  514661.1  301904.0  292465.2
#> [29]  265441.5  332884.0  295447.3  470034.4  439949.1  417708.1  413964.9
#> [36]  998798.1  304372.2  835378.4  333026.5  497092.9  304377.6  341107.2
#> [43]  497538.8  436456.4  728303.3  420439.9
#> 
#> $ranger_RS_8
#>  [1]  892866.2  386492.3  362563.7  704331.2  453090.5 1102717.2  322777.4
#>  [8]  507332.9  338399.7  343503.4  878342.7  387937.1  295821.8  476279.9
#> [15]  697134.9  731370.5  492933.2  417552.7  492012.2  484767.4  511067.8
#> [22]  514150.3  479579.6  300420.9  823106.9  505442.3  333765.0  321189.9
#> [29]  303379.6  352229.2  310429.8  559950.7  464120.1  441803.9  443126.9
#> [36]  923184.4  338257.2  864650.7  349603.3  573662.4  327956.3  360844.5
#> [43]  476279.9  471764.7  779717.5  432278.3
#> 
#> $ranger_RS_9
#>  [1]  962248.1  404771.3  330340.0  699198.7  467284.7 1002511.6  353646.7
#>  [8]  526374.2  338219.7  345140.3  852709.6  390296.9  358381.1  519341.3
#> [15]  715954.5  688644.7  520343.2  480682.1  495794.6  462500.0  515523.8
#> [22]  561718.1  522294.6  360280.1  733703.3  550329.4  381394.8  338089.3
#> [29]  366103.4  379800.4  338219.7  508475.8  467124.3  494195.7  491498.8
#> [36]  940965.2  397933.3  896214.5  417180.9  573673.1  352508.9  391196.5
#> [43]  519341.3  487542.3  760540.2  485503.4
#> 
#> $ranger_RS_10
#>  [1]  985291.7  453222.0  406528.1  735009.4  486069.5 1012724.4  379365.2
#>  [8]  563724.9  387664.6  399713.3  938757.2  448906.9  358143.3  534499.8
#> [15]  753593.9  805538.8  554264.5  455937.9  528506.9  523934.1  554533.3
#> [22]  561417.4  549201.0  358617.0  808066.2  556603.7  381621.4  374628.4
#> [29]  375203.8  406569.9  381808.2  586534.3  516263.0  507102.6  505305.7
#> [36]  954637.1  391090.0  865874.4  395818.9  563118.7  387188.5  398150.7
#> [43]  534499.8  509349.2  875387.4  474634.7
#> 
#> $xgboost_RS_1
#>  [1] 160320.69  43485.59  43485.59 124421.69  88234.16 160320.69  43485.59
#>  [8]  88234.16  43485.59  43485.59 160320.69  43485.59  43485.59  88234.16
#> [15] 124421.69 124421.69  88234.16  88234.16  88234.16  88234.16  88234.16
#> [22]  88234.16  88234.16  43485.59 160320.69  88234.16  43485.59  43485.59
#> [29]  43485.59  43485.59  43485.59  88234.16  88234.16  88234.16  88234.16
#> [36] 160320.69  43485.59 160320.69  43485.59  88234.16  43485.59  43485.59
#> [43]  88234.16  88234.16 124421.69  88234.16
#> 
#> $xgboost_RS_2
#>  [1] 452354.0 164056.7 218567.8 462079.2 275510.3 597598.5 164056.7 366803.6
#>  [9] 164056.7 218567.8 597598.5 164056.7 164056.7 366803.6 462079.2 370785.9
#> [17] 366803.6 275510.3 275510.3 275510.3 275510.3 366803.6 366803.6 164056.7
#> [25] 597598.5 366803.6 164056.7 164056.7 164056.7 164056.7 164056.7 370785.9
#> [33] 275510.3 275510.3 275510.3 597598.5 186698.1 597598.5 186698.1 370785.9
#> [41] 164056.7 218567.8 366803.6 275510.3 316834.7 275510.3
#> 
#> $xgboost_RS_3
#>  [1] 430398.6 117136.0 117136.0 322701.6 236993.5 430398.6 117136.0 236993.5
#>  [9] 117136.0 117136.0 430398.6 117136.0 117136.0 236993.5 322701.6 322701.6
#> [17] 236993.5 236993.5 236993.5 236993.5 236993.5 236993.5 236993.5 117136.0
#> [25] 430398.6 236993.5 117136.0 117136.0 117136.0 117136.0 117136.0 236993.5
#> [33] 236993.5 236993.5 236993.5 430398.6 117136.0 430398.6 117136.0 236993.5
#> [41] 117136.0 117136.0 236993.5 236993.5 322701.6 236993.5
#> 
#> $xgboost_RS_4
#>  [1] 426509.0 140553.2 140553.2 345197.4 189764.4 300314.1 140553.2 244540.3
#>  [9] 140553.2 140553.2 426509.0 140553.2 108273.1 244540.3 345197.4 290421.4
#> [17] 244540.3 189764.4 189764.4 189764.4 189764.4 244540.3 244540.3 108273.1
#> [25] 426509.0 244540.3 140553.2 140553.2 108273.1 140553.2 140553.2 164226.5
#> [33] 189764.4 189764.4 189764.4 426509.0 140553.2 426509.0 140553.2 290421.4
#> [41] 140553.2 140553.2 244540.3 189764.4 345197.4 189764.4
#> 
#> $xgboost_RS_5
#>  [1] 426509.0 140553.2 140553.2 345197.4 189764.4 300314.1 140553.2 244540.3
#>  [9] 140553.2 140553.2 426509.0 140553.2 108273.1 244540.3 345197.4 290421.4
#> [17] 244540.3 189764.4 189764.4 189764.4 189764.4 244540.3 244540.3 108273.1
#> [25] 426509.0 244540.3 140553.2 140553.2 108273.1 140553.2 140553.2 164226.5
#> [33] 189764.4 189764.4 189764.4 426509.0 140553.2 426509.0 140553.2 290421.4
#> [41] 140553.2 140553.2 244540.3 189764.4 345197.4 189764.4
#> 
#> $xgboost_RS_6
#>  [1] 629176.1 197482.2 197482.2 449681.1 320196.1 629176.1 197482.2 320196.1
#>  [9] 197482.2 197482.2 629176.1 197482.2 197482.2 320196.1 449681.1 449681.1
#> [17] 320196.1 320196.1 320196.1 320196.1 320196.1 320196.1 320196.1 197482.2
#> [25] 629176.1 320196.1 197482.2 197482.2 197482.2 197482.2 197482.2 449681.1
#> [33] 320196.1 320196.1 320196.1 629176.1 197482.2 629176.1 197482.2 449681.1
#> [41] 197482.2 197482.2 320196.1 320196.1 449681.1 320196.1
#> 
#> $xgboost_RS_7
#>  [1] 452354.0 164056.7 218567.8 462079.2 275510.3 597598.5 164056.7 366803.6
#>  [9] 164056.7 218567.8 597598.5 164056.7 164056.7 366803.6 462079.2 370785.9
#> [17] 366803.6 275510.3 275510.3 275510.3 275510.3 366803.6 366803.6 164056.7
#> [25] 597598.5 366803.6 164056.7 164056.7 164056.7 164056.7 164056.7 370785.9
#> [33] 275510.3 275510.3 275510.3 597598.5 186698.1 597598.5 186698.1 370785.9
#> [41] 164056.7 218567.8 366803.6 275510.3 316834.7 275510.3
#> 
#> $xgboost_RS_8
#>  [1] 452354.0 164056.7 217018.8 462079.2 273961.3 597598.5 164056.7 365254.6
#>  [9] 164056.7 217018.8 597598.5 164056.7 164056.7 365254.6 462079.2 370785.9
#> [17] 365254.6 273961.3 273961.3 273961.3 273961.3 365254.6 365254.6 164056.7
#> [25] 597598.5 365254.6 164056.7 164056.7 164056.7 164056.7 164056.7 370785.9
#> [33] 273961.3 273961.3 273961.3 597598.5 186698.1 597598.5 186698.1 370785.9
#> [41] 164056.7 217018.8 365254.6 273961.3 316834.7 273961.3
#> 
#> $xgboost_RS_9
#>  [1] 426509.0 140553.2 140553.2 345197.4 189764.4 300314.1 140553.2 244540.3
#>  [9] 140553.2 140553.2 426509.0 140553.2 108273.1 244540.3 345197.4 290421.4
#> [17] 244540.3 189764.4 189764.4 189764.4 189764.4 244540.3 244540.3 108273.1
#> [25] 426509.0 244540.3 140553.2 140553.2 108273.1 140553.2 140553.2 164226.5
#> [33] 189764.4 189764.4 189764.4 426509.0 140553.2 426509.0 140553.2 290421.4
#> [41] 140553.2 140553.2 244540.3 189764.4 345197.4 189764.4
#> 
#> $xgboost_RS_10
#>  [1] 163847.75  51002.20  51002.20 107029.89  72243.28 163847.75  51002.20
#>  [8] 107029.89  51002.20  51002.20 163847.75  51002.20  33266.60 107029.89
#> [15] 107029.89  72243.28 107029.89  72243.28  72243.28  72243.28  72243.28
#> [22] 107029.89 107029.89  33266.60 163847.75 107029.89  51002.20  51002.20
#> [29]  33266.60  51002.20  51002.20  72243.28  72243.28  72243.28  72243.28
#> [36] 163847.75  51002.20 163847.75  51002.20  72243.28  51002.20  51002.20
#> [43] 107029.89  72243.28 107029.89  72243.28
#> 
#> $decision_tree_RS_1
#>  [1] 1049571.4  309000.0  309000.0 1049571.4  395181.8  722674.2  195018.2
#>  [8]  395181.8  195018.2  195018.2  722674.2  309000.0  195018.2  560642.9
#> [15] 1049571.4  722674.2  395181.8  395181.8  395181.8  560642.9  560642.9
#> [22]  395181.8  560642.9  195018.2  722674.2  560642.9  195018.2  195018.2
#> [29]  195018.2  309000.0  195018.2  560642.9  395181.8  395181.8  560642.9
#> [36]  912500.0  195018.2  722674.2  195018.2  395181.8  309000.0  309000.0
#> [43]  560642.9  395181.8 1049571.4  560642.9
#> 
#> $decision_tree_RS_2
#>  [1] 1049571.4  195018.2  195018.2 1049571.4  560642.9  722674.2  309000.0
#>  [8]  395181.8  195018.2  195018.2  722674.2  309000.0  309000.0  395181.8
#> [15] 1049571.4  722674.2  395181.8  395181.8  395181.8  395181.8  560642.9
#> [22]  395181.8  560642.9  195018.2  722674.2  395181.8  195018.2  195018.2
#> [29]  309000.0  309000.0  309000.0  560642.9  395181.8  560642.9  395181.8
#> [36]  912500.0  195018.2  722674.2  309000.0  395181.8  309000.0  195018.2
#> [43]  560642.9  395181.8 1049571.4  560642.9
#> 
#> $decision_tree_RS_3
#>  [1] 1049571.4  195018.2  195018.2 1049571.4  395181.8  722674.2  195018.2
#>  [8]  395181.8  195018.2  195018.2  722674.2  195018.2  309000.0  560642.9
#> [15] 1049571.4  722674.2  560642.9  395181.8  395181.8  395181.8  395181.8
#> [22]  395181.8  395181.8  195018.2  722674.2  395181.8  195018.2  309000.0
#> [29]  195018.2  195018.2  195018.2  395181.8  395181.8  560642.9  395181.8
#> [36]  912500.0  195018.2  722674.2  309000.0  395181.8  195018.2  309000.0
#> [43]  395181.8  560642.9 1049571.4  560642.9
#> 
#> $decision_tree_RS_4
#>  [1] 1049571.4  195018.2  195018.2 1049571.4  395181.8  722674.2  195018.2
#>  [8]  560642.9  309000.0  309000.0  722674.2  195018.2  195018.2  560642.9
#> [15] 1049571.4  722674.2  560642.9  395181.8  395181.8  560642.9  560642.9
#> [22]  395181.8  395181.8  195018.2  722674.2  395181.8  309000.0  195018.2
#> [29]  195018.2  309000.0  195018.2  395181.8  395181.8  395181.8  560642.9
#> [36]  912500.0  195018.2  722674.2  195018.2  560642.9  309000.0  195018.2
#> [43]  395181.8  560642.9 1049571.4  395181.8
#> 
#> $decision_tree_RS_5
#>  [1] 1049571.4  195018.2  195018.2 1049571.4  560642.9  722674.2  309000.0
#>  [8]  395181.8  195018.2  195018.2  722674.2  309000.0  195018.2  395181.8
#> [15] 1049571.4  722674.2  395181.8  560642.9  395181.8  395181.8  560642.9
#> [22]  395181.8  560642.9  309000.0  722674.2  395181.8  195018.2  309000.0
#> [29]  309000.0  309000.0  195018.2  395181.8  395181.8  395181.8  395181.8
#> [36]  912500.0  309000.0  722674.2  309000.0  560642.9  309000.0  309000.0
#> [43]  560642.9  560642.9 1049571.4  560642.9
#> 
#> $decision_tree_RS_6
#>  [1] 1049571.4  309000.0  195018.2 1049571.4  395181.8  722674.2  195018.2
#>  [8]  395181.8  195018.2  195018.2  722674.2  195018.2  195018.2  560642.9
#> [15] 1049571.4  722674.2  395181.8  560642.9  395181.8  395181.8  395181.8
#> [22]  395181.8  560642.9  195018.2  722674.2  395181.8  195018.2  309000.0
#> [29]  309000.0  309000.0  195018.2  395181.8  395181.8  395181.8  560642.9
#> [36]  912500.0  195018.2  722674.2  309000.0  395181.8  195018.2  309000.0
#> [43]  395181.8  395181.8 1049571.4  395181.8
#> 
#> $decision_tree_RS_7
#>  [1] 1049571.4  309000.0  195018.2 1049571.4  560642.9  722674.2  195018.2
#>  [8]  560642.9  195018.2  309000.0  722674.2  195018.2  195018.2  560642.9
#> [15] 1049571.4  722674.2  395181.8  395181.8  395181.8  560642.9  395181.8
#> [22]  560642.9  395181.8  309000.0  722674.2  560642.9  195018.2  195018.2
#> [29]  195018.2  195018.2  195018.2  560642.9  560642.9  395181.8  395181.8
#> [36]  912500.0  195018.2  722674.2  195018.2  560642.9  195018.2  195018.2
#> [43]  560642.9  560642.9 1049571.4  560642.9
#> 
#> $decision_tree_RS_8
#>  [1] 1049571.4  309000.0  195018.2 1049571.4  395181.8  722674.2  195018.2
#>  [8]  395181.8  195018.2  195018.2  722674.2  195018.2  195018.2  395181.8
#> [15] 1049571.4  722674.2  560642.9  395181.8  395181.8  395181.8  395181.8
#> [22]  395181.8  560642.9  195018.2  722674.2  395181.8  309000.0  309000.0
#> [29]  195018.2  195018.2  309000.0  395181.8  560642.9  395181.8  560642.9
#> [36]  912500.0  309000.0  722674.2  195018.2  395181.8  195018.2  309000.0
#> [43]  560642.9  560642.9 1049571.4  560642.9
#> 
#> $decision_tree_RS_9
#>  [1] 1049571.4  309000.0  195018.2 1049571.4  395181.8  722674.2  195018.2
#>  [8]  395181.8  309000.0  195018.2  722674.2  195018.2  195018.2  560642.9
#> [15] 1049571.4  722674.2  395181.8  560642.9  560642.9  395181.8  395181.8
#> [22]  395181.8  395181.8  309000.0  722674.2  395181.8  195018.2  195018.2
#> [29]  195018.2  309000.0  195018.2  560642.9  560642.9  395181.8  560642.9
#> [36]  912500.0  195018.2  722674.2  195018.2  395181.8  195018.2  195018.2
#> [43]  560642.9  560642.9 1049571.4  395181.8
#> 
#> $decision_tree_RS_10
#>  [1] 1049571.4  195018.2  309000.0 1049571.4  395181.8  722674.2  309000.0
#>  [8]  395181.8  195018.2  195018.2  722674.2  195018.2  309000.0  395181.8
#> [15] 1049571.4  722674.2  395181.8  560642.9  560642.9  395181.8  560642.9
#> [22]  560642.9  395181.8  195018.2  722674.2  395181.8  309000.0  309000.0
#> [29]  195018.2  195018.2  309000.0  560642.9  395181.8  395181.8  560642.9
#> [36]  912500.0  309000.0  722674.2  195018.2  395181.8  195018.2  309000.0
#> [43]  395181.8  395181.8 1049571.4  395181.8
#> 
#> $lightgbm_RS_1
#>  [1]  593943.4  251562.2  225028.0  704192.8  295634.0  565289.9  117118.9
#>  [8]  661816.5  173984.4  232505.0 1197920.4  245408.2  155916.0  430531.0
#> [15]  704192.8  255269.9  548471.7  214119.0  387379.8  383299.9  349062.7
#> [22]  639250.4  493782.3  234487.1  522579.9  537410.7  258245.8  197433.2
#> [29]  251948.9  238004.0  135363.0  389185.0  389185.0  355065.9  355065.9
#> [36] 2859077.8  224740.4  818411.8  332624.5  240434.0  311769.2  141705.4
#> [43]  430531.0  370258.8  257503.0  187964.2
#> 
#> $lightgbm_RS_2
#>  [1]  708227.4  191288.9  208343.4  763704.9  309768.9  613737.6  109721.2
#>  [8]  624700.7  176714.7  224250.9 1013972.5  172747.7  231328.0  423478.4
#> [15]  763704.9  376900.4  590029.2  219598.6  322599.8  361474.5  302807.3
#> [22]  630140.9  499312.7  263264.3  579407.7  536757.6  181128.3  161943.1
#> [29]  270241.1  191288.9  169972.4  368415.2  368415.2  368415.2  363515.9
#> [36] 1885881.6  230020.4  856542.3  323197.4  277483.8  241432.1  164627.4
#> [43]  423478.4  322599.8  441682.2  225511.4
#> 
#> $lightgbm_RS_3
#>  [1]  673698.0  200819.2  206904.4  756843.5  308255.2  557293.7  112051.3
#>  [8]  611360.4  178876.3  239192.4 1034830.1  190512.6  219092.4  435365.1
#> [15]  756843.5  325974.6  586875.6  214458.3  332585.2  366494.8  313859.4
#> [22]  627096.6  504108.2  261992.2  526492.2  551091.0  215931.4  190423.0
#> [29]  274753.4  200819.2  164274.8  364693.3  364693.3  364693.3  364693.3
#> [36] 2053117.2  227717.3  809633.7  336197.3  253580.8  279963.2  149559.2
#> [43]  435365.1  332585.2  380840.6  224425.2
#> 
#> $lightgbm_RS_4
#>  [1]  837467.5  221635.1  242342.1  786822.1  345635.5  684462.7  215198.9
#>  [8]  643326.9  234704.7  232035.5  972776.0  211328.5  301544.4  426260.6
#> [15]  786822.1  495919.3  653635.5  310321.4  349665.6  345635.5  339359.0
#> [22]  620389.1  472791.0  309433.4  676241.8  523306.8  260429.6  226703.7
#> [29]  322194.6  221635.1  220555.6  377569.7  377569.7  377569.7  377569.7
#> [36] 1388727.7  235430.2  916408.5  308413.1  306947.7  263969.6  198242.2
#> [43]  426260.6  349665.6  560380.9  310321.4
#> 
#> $lightgbm_RS_5
#>  [1] 778229.9 385084.2 402287.3 643751.1 490335.0 742375.3 385084.2 571665.0
#>  [9] 385084.2 385084.2 754884.2 385084.2 374745.3 533359.2 643751.1 603175.0
#> [17] 566765.7 469437.4 490335.0 490335.0 495234.3 571665.0 566765.7 374745.3
#> [25] 742375.3 566765.7 385084.2 385084.2 374745.3 385084.2 385084.2 495234.3
#> [33] 495234.3 495234.3 490335.0 986319.1 374745.3 754884.2 385084.2 495234.3
#> [41] 385084.2 402287.3 533359.2 490335.0 603175.0 469437.4
#> 
#> $lightgbm_RS_6
#>  [1]  761311.1  209011.3  207970.5  781287.7  307859.0  639103.7  141062.3
#>  [8]  627511.6  190961.5  221635.9  950389.1  189839.6  253381.4  408062.8
#> [15]  781287.7  410452.3  601289.6  229680.8  313442.5  335200.5  295204.0
#> [22]  608500.4  486414.6  292773.0  610240.2  525420.6  186528.6  161682.3
#> [29]  299220.2  209011.3  174701.3  358725.9  358725.9  358725.9  358725.9
#> [36] 1670111.3  218043.9  865279.7  317952.9  267193.5  238572.8  161829.2
#> [43]  408062.8  313442.5  507112.1  235102.7
#> 
#> $lightgbm_RS_7
#>  [1]  761311.1  209011.3  207970.5  781287.7  307859.0  639103.7  141062.3
#>  [8]  627511.6  190961.5  221635.9  950389.1  189839.6  253381.4  408062.8
#> [15]  781287.7  410452.3  601289.6  229680.8  313442.5  335200.5  295204.0
#> [22]  608500.4  486414.6  292773.0  610240.2  525420.6  186528.6  161682.3
#> [29]  299220.2  209011.3  174701.3  358725.9  358725.9  358725.9  358725.9
#> [36] 1670111.3  218043.9  865279.7  317952.9  267193.5  238572.8  161829.2
#> [43]  408062.8  313442.5  507112.1  235102.7
#> 
#> $lightgbm_RS_8
#>  [1]  803291.5  218415.3  248472.1  778129.7  340160.9  643369.3  201774.2
#>  [8]  641849.8  218938.9  214823.2  897644.3  201969.6  288172.2  403796.4
#> [15]  778129.7  485606.6  645322.3  299819.5  343753.0  336933.7  328979.4
#> [22]  634600.5  458650.9  320108.6  644230.8  512108.5  263809.3  197764.8
#> [29]  327085.4  218415.3  205800.6  359400.0  359400.0  359400.0  354500.7
#> [36] 1404367.8  220373.4  887064.9  298548.2  296904.4  229163.0  187097.8
#> [43]  403796.4  343753.0  576653.7  296592.3
#> 
#> $lightgbm_RS_9
#>  [1] 778229.9 385084.2 402287.3 643751.1 490335.0 742375.3 385084.2 571665.0
#>  [9] 385084.2 385084.2 754884.2 385084.2 374745.3 533359.2 643751.1 603175.0
#> [17] 566765.7 469437.4 490335.0 490335.0 495234.3 571665.0 566765.7 374745.3
#> [25] 742375.3 566765.7 385084.2 385084.2 374745.3 385084.2 385084.2 495234.3
#> [33] 495234.3 495234.3 490335.0 986319.1 374745.3 754884.2 385084.2 495234.3
#> [41] 385084.2 402287.3 533359.2 490335.0 603175.0 469437.4
#> 
#> $lightgbm_RS_10
#>  [1]  837467.5  221635.1  242342.1  786822.1  345635.5  684462.7  215198.9
#>  [8]  643326.9  234704.7  232035.5  972776.0  211328.5  301544.4  426260.6
#> [15]  786822.1  495919.3  653635.5  310321.4  349665.6  345635.5  339359.0
#> [22]  620389.1  472791.0  309433.4  676241.8  523306.8  260429.6  226703.7
#> [29]  322194.6  221635.1  220555.6  377569.7  377569.7  377569.7  377569.7
#> [36] 1388727.7  235430.2  916408.5  308413.1  306947.7  263969.6  198242.2
#> [43]  426260.6  349665.6  560380.9  310321.4
#> 
#> $ranger_bayes
#>  [1] 883808.1 357646.3 342488.7 678370.1 434316.7 988377.5 318518.3 480990.9
#>  [9] 321720.9 333028.8 909177.5 359983.8 306973.3 498138.4 711139.4 717590.3
#> [17] 489344.6 409023.3 445127.1 440294.4 495977.4 485579.0 512590.6 303460.4
#> [25] 771706.5 516847.9 325489.8 307275.4 301436.9 335901.4 311184.7 474840.5
#> [33] 431455.4 421081.7 425270.2 975128.1 367614.7 815982.9 366146.6 491297.3
#> [41] 328473.3 332476.9 498138.4 462710.1 779501.2 454105.5
#> 
#> $xgboost_bayes
#>  [1]  705548.7  299250.2  252848.3  682319.5  294562.4  638772.8  267457.0
#>  [8]  558879.1  233177.5  272977.4 1139620.1  259021.1  243005.1  525752.3
#> [15]  908681.2  355873.1  469589.9  307976.5  350621.3  283477.6  480625.7
#> [22]  588061.8  653536.5  230647.7  597018.9  552706.2  203937.7  215367.1
#> [29]  168924.6  228126.1  205987.7  432455.3  336445.5  336445.5  336445.5
#> [36] 1385546.6  349792.1  802391.1  401300.3  287542.4  258021.5  246165.4
#> [43]  525752.3  470309.4  674504.4  392735.0
#> 
#> $decision_tree_bayes
#>  [1] 1055875.0  309000.0  195018.2  445823.5  445823.5  729436.4  195018.2
#>  [8]  445823.5  309000.0  309000.0 1055875.0  195018.2  195018.2  445823.5
#> [15]  445823.5  445823.5  445823.5  445823.5  445823.5  445823.5  445823.5
#> [22]  445823.5  445823.5  195018.2  729436.4  445823.5  195018.2  309000.0
#> [29]  195018.2  195018.2  309000.0  445823.5  445823.5  445823.5  445823.5
#> [36]  912500.0  309000.0  729436.4  195018.2  445823.5  309000.0  195018.2
#> [43]  445823.5  445823.5  445823.5  445823.5
#> 
#> $lightgbm_bayes
#>  [1]  744194.6  187683.9  213667.4  771566.2  314858.4  628322.8  114897.9
#>  [8]  613825.8  195332.0  222044.2  989409.1  175379.7  227250.9  410680.7
#> [15]  771566.2  385735.6  576572.2  224600.9  308248.2  346900.2  297064.3
#> [22]  605626.9  479245.6  270777.0  599652.8  510752.6  165644.8  160303.6
#> [29]  279268.7  187683.9  175482.1  387829.1  387829.1  387829.1  380266.2
#> [36] 1757579.5  212367.2  855068.8  316487.5  259566.8  245358.3  152404.9
#> [43]  410680.7  308248.2  492070.2  228892.7
#>