Train models with forester

The `train()` function is the core function of this package. The only obligatory arguments are `data` and `target`. Setting and changing other arguments will affect model validation strategy, tested model families, and so on.

train(
  data,
  y,
  type = "guess",
  engine = c("ranger", "xgboost", "decision_tree", "lightgbm", "catboost"),
  loss = "default",
  validation = "default",
  tuning = "default",
  keep = FALSE,
  verbose = TRUE,
  bayes_iter = 10,
  random_iter = 10,
  advanced_preprocessing = FALSE
)

Arguments

data

A `data.frame` or `matrix` - data which will be used to build models. By default model will be trained on all columns in the `data`.

y

A target variable. It can be either (1) a vector of the same number of observations as `data` or (2) a character name of variable in the `data` that contains the target variable.

type

A character, one of `classification`/`regression`/`guess` that sets the type of the task. If `guess` (the default option) then forester will figure out `type` based on the number of unique values in the `y` variable.

engine

A vector of tree-based models that shall be tested. Possible values are: `ranger`, `xgboost`, `decision_tree`, `lightgbm`, `catboost`. All models from this vector will be trained and the best one will be returned.

loss

A character with the name of the loss function. If `default` then the default loss function for a given task will be used.

validation

A validation/model selection a character of data.frame. If `data.frame` then this data will be used for the model assessment and selection. If `character`, then it will be a name of the hyperparameter.

tuning

A character with the name of the tuning procedure. one out of `default`/`portfolio`/`bayesian`/`random search`. If `default` then no tuning, just default hyperparameteres are used. If `portfolio` then static portfolio of hyperparametes are tried. If `bayesian` then Bayesian optimization is performed. If `random search` then random search is performed.

keep

Shall all models be returned (`keep = FALSE`) or only the best one (`keep = FALSE`, default option)?

verbose

A logical value, if set to TRUE, provides all information about training process, if FALSE gives none.

bayes_iter

An integer value describing number of optimization rounds used by Bayesian optimization.

random_iter

An integer value describing number of optimization rounds used by random search.

advanced_preprocessing

A logical value describing, whether the user wants to use advanced preprocessing methods (ex. deleting correlated values).

Value

A list of all necessary objects for other functions.

Examples

library(forester)
data('lisbon')
train_output <- train(lisbon, 'Price')
#> Type guessed as:  regression 
#>  -------------------- CHECK DATA REPORT -------------------- 
#>  
#> The dataset has 246 observations and 17 columns, which names are: 
#> Id; Condition; PropertyType; PropertySubType; Bedrooms; Bathrooms; AreaNet; AreaGross; Parking; Latitude; Longitude; Country; District; Municipality; Parish; Price.M2; Price; 
#> 
#> With the target value described by a column Price.
#> 
#> Static columns are: 
#> Country; District; Municipality; 
#> 
#> With dominating values: 
#> Portugal; Lisboa; Lisboa; 
#> 
#> These column pairs are duplicate:
#>  District - Municipality; 
#> 
#> No target values are missing. 
#> 
#> No predictor values are missing. 
#> 
#> Warning: Number of logged events: 4
#> No issues with dimensionality. 
#> 
#> Strongly correlated pairs of numerical values are: 
#>  
#> Bedrooms - AreaNet: 0.77;
#> Bedrooms - AreaGross: 0.77;
#> Bathrooms - AreaNet: 0.78;
#> Bathrooms - AreaGross: 0.78;
#> AreaNet - AreaGross: 1;
#> 
#> Strongly correlated pairs of categorical values are: 
#> PropertyType - PropertySubType: 1;
#> 
#> These obserwation migth be outliers due to their numerical columns values: 
#>  145 146 196 44 5 51 57 58 59 60 61 62 63 64 69 75 76 77 78 ;
#> 
#> Target data is not evenly distributed with quantile bins: 0.25 0.35 0.14 0.26 
#> 
#> Columns names suggest that some of them are IDs, removing them can improve the model.
#>  Suspicious columns are: Id .
#> 
#> Columns data suggest that some of them are IDs, removing them can improve the model.
#>  Suspicious columns are: Id .
#> 
#>  -------------------- CHECK DATA REPORT END -------------------- 
#>  
#> Warning: Number of logged events: 1
#> Data preprocessed. 
#> Data splitted and balanced 
#> Warning: NAs introduced by coercion
#> Correct formats prepared 
#> Models sucsesfully trained 
#> Warning: NAs introduced by coercion
#> Predicted Successfully 
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> 
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
train_output$ranked_list
#>                 engine          mse     rmse          r2       mad       mae
#> 1        catboost_RS_7  13287089514 115269.6  0.88480199  44691.14  81219.87
#> 2        catboost_RS_5  16322628450 127760.0  0.85848410  65281.45  91677.04
#> 3        catboost_RS_6  17316210005 131591.1  0.84986983  57494.50  95512.92
#> 4        catboost_RS_1  17601989487 132672.5  0.84739214  73947.95  99199.62
#> 5       catboost_RS_10  18432150588 135765.1  0.84019471  56550.68  91548.26
#> 6        xgboost_model  20199334629 142124.4  0.82487337  69611.20 104424.18
#> 7        xgboost_bayes  20199334629 142124.4  0.82487337  69611.20 104424.18
#> 8       catboost_model  22255799235 149183.8  0.80704398  57172.77 100257.28
#> 9       catboost_bayes  22255799235 149183.8  0.80704398  57172.77 100257.28
#> 10       catboost_RS_2  22778811778 150926.5  0.80250950  77649.90 104368.69
#> 11         ranger_RS_4  23515770858 153348.5  0.79612013  73526.68 109138.60
#> 12        ranger_model  23526276072 153382.8  0.79602905  79783.93 109855.88
#> 13        ranger_bayes  23526276072 153382.8  0.79602905  79783.93 109855.88
#> 14         ranger_RS_1  24830309248 157576.4  0.78472318  81652.33 113563.22
#> 15         ranger_RS_9  25774245592 160543.6  0.77653933  93215.43 120099.65
#> 16         ranger_RS_5  26018600366 161302.8  0.77442079  91039.68 113483.59
#> 17       lightgbm_RS_3  26682593880 163348.1  0.76866402 104657.35 131276.74
#> 18         ranger_RS_2  29132575865 170682.7  0.74742287  87335.46 125369.17
#> 19       lightgbm_RS_8  30054948172 173363.6  0.73942598 130717.74 137987.51
#> 20       lightgbm_RS_1  31001532149 176072.5  0.73121917 137677.28 142193.73
#> 21       lightgbm_RS_9  31001532149 176072.5  0.73121917 137677.28 142193.73
#> 22         ranger_RS_3  31931972171 178695.2  0.72315233  80259.62 129696.00
#> 23       lightgbm_RS_5  35669580105 188863.9  0.69074756 143468.69 154934.01
#> 24       lightgbm_RS_7  35669580105 188863.9  0.69074756 143468.69 154934.01
#> 25      lightgbm_model  38886793073 197197.3  0.66285458 151084.45 158269.20
#> 26      lightgbm_RS_10  38886793073 197197.3  0.66285458 151084.45 158269.20
#> 27      lightgbm_bayes  38886793073 197197.3  0.66285458 151084.45 158269.20
#> 28         ranger_RS_7  48752122419 220798.8  0.57732296 143296.59 173315.02
#> 29         ranger_RS_8  48867653236 221060.3  0.57632131 159712.87 177752.09
#> 30       lightgbm_RS_4  52484461458 229094.9  0.54496387 146944.08 170433.07
#> 31        ranger_RS_10  53555515962 231420.6  0.53567791 143437.27 174632.45
#> 32         ranger_RS_6  55895702143 236422.7  0.51538869 176991.20 184704.21
#> 33        xgboost_RS_4  58372530954 241604.1  0.49391478 110344.26 170303.54
#> 34        xgboost_RS_2  59025904869 242952.5  0.48825008 121403.32 173060.61
#> 35        xgboost_RS_8  59025904869 242952.5  0.48825008 121403.32 173060.61
#> 36       lightgbm_RS_6  60916666125 246813.0  0.47185733 155198.66 178300.46
#> 37        xgboost_RS_1  61969483284 248936.7  0.46272948 105078.35 167997.57
#> 38       catboost_RS_3  74351717001 272675.1  0.35537650 114631.39 195331.85
#> 39       lightgbm_RS_2  82059260120 286459.9  0.28855271 156728.04 191050.06
#> 40 decision_tree_model  87381240462 295603.2  0.24241156 107316.33 192250.00
#> 41  decision_tree_RS_1  87381240462 295603.2  0.24241156 107316.33 192250.00
#> 42  decision_tree_RS_2  87381240462 295603.2  0.24241156 107316.33 192250.00
#> 43  decision_tree_RS_3  87381240462 295603.2  0.24241156 107316.33 192250.00
#> 44  decision_tree_RS_4  87381240462 295603.2  0.24241156 107316.33 192250.00
#> 45  decision_tree_RS_5  87381240462 295603.2  0.24241156 107316.33 192250.00
#> 46  decision_tree_RS_6  87381240462 295603.2  0.24241156 107316.33 192250.00
#> 47  decision_tree_RS_7  87381240462 295603.2  0.24241156 107316.33 192250.00
#> 48  decision_tree_RS_8  87381240462 295603.2  0.24241156 107316.33 192250.00
#> 49  decision_tree_RS_9  87381240462 295603.2  0.24241156 107316.33 192250.00
#> 50 decision_tree_RS_10  87381240462 295603.2  0.24241156 107316.33 192250.00
#> 51 decision_tree_bayes  87381240462 295603.2  0.24241156 107316.33 192250.00
#> 52        xgboost_RS_9 126704488007 355955.7 -0.09851788 217096.28 271146.86
#> 53        xgboost_RS_5 126799535593 356089.2 -0.09934194 217096.28 271380.98
#> 54        xgboost_RS_7 126799535593 356089.2 -0.09934194 217096.28 271380.98
#> 55       catboost_RS_8 154020882531 392454.9 -0.33534886 117203.10 248219.41
#> 56       catboost_RS_4 182438090265 427127.7 -0.58172381 139678.59 240839.03
#> 57       catboost_RS_9 193383321428 439753.7 -0.67661809 136302.98 253004.71
#> 58        xgboost_RS_6 274563883507 523988.4 -1.38044714 342247.48 427370.89
#> 59        xgboost_RS_3 274677664401 524097.0 -1.38143361 344352.06 430301.77
#> 60       xgboost_RS_10 274677664401 524097.0 -1.38143361 344352.06 430301.77